pyproject.toml
[tool.poetry]
name = "scale-img"
version = "0.1.0"
description = ""
authors = ["HotariTobu <hotari24tools@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pillow = "^11.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
